graphics-state-line-style-border: specialize on medium
line-style-effective-thickness is expected to be called on a medium,
but its current stub implementation masks the issue by not checking
for the argument type. The issue is that g-s-l-b could have called
l-s-e-t with a sheet instead of that sheet medium. This commit adds a
specialization for g-s-l-b on a sheet which calls l-s-e-t with its
medium.